refactor: Repository abstract에서 interface로 수정
